sql_agent/pyproject.toml
[project]
name = "sql-agent"
version = "0.1.0"
description = "An AI-powered SQL data analyst agent built with Claude and the Anthropic Python SDK"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.84.0",
    "pandas>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
sql-agent = "main:main"
